# contracts/openapi/sentinelops-api.yaml
openapi: 3.1.0
info:
  title: SentinelOps AI Stage 1 API
  version: 0.1.0-demo
  description: >-
    Service-scoped incident command API. Mutating operations are explicit,
    idempotent, version-checked, role-gated, and auditable.
servers:
  - url: /api/v1
security:
  - bearerAuth: []
...
paths:
  /incidents:
    get:
      operationId: listIncidents
      tags: [Incidents]
      summary: List service-scoped incidents
      x-required-roles: [OBSERVER, ON_CALL_OPERATOR, SRE_APPROVER, RUNBOOK_ADMIN, PLATFORM_ADMIN]
      parameters:
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/IncidentStatus"
        - name: serviceId
          in: query
          schema:
            type: string
            format: uuid
        - name: severity
          in: query
          schema:
            $ref: "#/components/schemas/Severity"
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: A cursor-paginated incident page.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IncidentPage"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /incidents/{incidentId}:
    parameters:
      - $ref: "#/components/parameters/IncidentId"
    get:
      operationId: getIncidentCockpit
      tags: [Incidents]
      summary: Get a sanitized incident cockpit projection
      description: >-
        Returns the latest bounded diagnosis, evidence metadata, active approval,
        and execution. Raw prompts, execution tickets, provider responses, and
        evidence bodies are deliberately excluded.
      x-required-roles: [OBSERVER, ON_CALL_OPERATOR, SRE_APPROVER, RUNBOOK_ADMIN, PLATFORM_ADMIN]
      responses:
        "200":
          description: Service-scoped incident cockpit.
          headers:
            ETag:
              $ref: "#/components/headers/ETag"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IncidentCockpit"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /incidents/{incidentId}/timeline:
    parameters:
      - $ref: "#/components/parameters/IncidentId"
    get:
      operationId: getIncidentTimeline
      tags: [Incidents]
      summary: Read the immutable incident event timeline
      x-required-roles: [OBSERVER, ON_CALL_OPERATOR, SRE_APPROVER, RUNBOOK_ADMIN, PLATFORM_ADMIN]
      parameters:
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: A cursor-paginated event page.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IncidentTimelinePage"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /incidents/{incidentId}/evidence:
    parameters:
      - $ref: "#/components/parameters/IncidentId"
    get:
      operationId: listIncidentEvidence
      tags: [Incidents]
      summary: Read separately authorized redacted evidence snapshots
      x-required-roles: [OBSERVER, ON_CALL_OPERATOR, SRE_APPROVER, RUNBOOK_ADMIN, PLATFORM_ADMIN]
      responses:
        "200":
          description: Redacted evidence snapshots.
          content:
            application/json:
              schema:
                type: array
                maxItems: 100
                items:
                  $ref: "#/components/schemas/EvidenceSnapshot"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IncidentId:
      name: incidentId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    Cursor:
      name: cursor
      in: query
      description: Opaque cursor returned by the preceding page.
      schema:
        type: string
    PageSize:
      name: pageSize
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 50
...
  headers:
    ETag:
      description: Quoted incident resource version for subsequent If-Match commands.
      schema:
        type: string
        pattern: '^"(0|[1-9][0-9]*)"$'
        example: '"7"'
  responses:
    BadRequest:
      description: The request is malformed or contains invalid values.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetail"
    Unauthorized:
      description: A valid bearer token is required.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetail"
    Forbidden:
      description: The principal lacks the role or service scope.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetail"
    NotFound:
      description: The resource does not exist.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetail"
...
  schemas:
...
    IncidentStatus:
      type: string
      enum:
        - DETECTED
        - TRIAGING
        - DIAGNOSED
        - AWAITING_APPROVAL
        - EXECUTING
        - VERIFYING
        - RESOLVED
        - SUPPRESSED
        - ESCALATED
    Severity:
      type: string
      enum: [sev1, sev2, sev3, sev4]
    RiskLevel:
      type: string
      enum: [R0, R1, R2, R3]
    ApprovalStatus:
      type: string
      enum: [PENDING, APPROVED, REJECTED, EXPIRED, INVALIDATED]
    ExecutionStatus:
      type: string
      enum: [PENDING, RUNNING, VERIFYING, SUCCEEDED, FAILED, UNKNOWN, ESCALATED]
...
    ProblemDetail:
      type: object
      required: [type, title, status, detail, instance, errorCode, traceId]
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
          minimum: 400
          maximum: 599
        detail:
          type: string
        instance:
          type: string
          format: uri-reference
        errorCode:
          type: string
        traceId:
          type: string
      additionalProperties: true
    IncidentSummary:
      type: object
      additionalProperties: false
      required:
        - id
        - serviceId
        - serviceKey
        - title
        - severity
        - status
        - version
        - occurrenceCount
        - openedAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
        serviceId:
          type: string
          format: uuid
        serviceKey:
          type: string
        title:
          type: string
        severity:
          $ref: "#/components/schemas/Severity"
        status:
          $ref: "#/components/schemas/IncidentStatus"
        version:
          type: integer
          format: int64
          minimum: 0
        occurrenceCount:
          type: integer
          format: int64
          minimum: 1
        openedAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        resolvedAt:
          type: [string, "null"]
          format: date-time
    IncidentPage:
      type: object
      additionalProperties: false
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/IncidentSummary"
        nextCursor:
          type: [string, "null"]
    Hypothesis:
      type: object
      additionalProperties: false
      required: [rank, statement, confidence, evidenceRefs]
      properties:
        rank:
          type: integer
          minimum: 1
        statement:
          type: string
        confidence:
          type: number
          minimum: 0
          maximum: 1
        evidenceRefs:
          type: array
          items:
            type: string
            format: uuid
    VerificationExpectation:
      type: object
      additionalProperties: false
      required: [probe, successThreshold, attempts, intervalSeconds]
      properties:
        probe:
          type: string
        successThreshold:
          type: number
        attempts:
          type: integer
          minimum: 1
        intervalSeconds:
          type: integer
          minimum: 0
...
    CockpitDiagnosis:
      type: object
      additionalProperties: false
      required:
        - id
        - diagnosisRunId
        - incidentVersion
        - summary
        - hypotheses
        - missingEvidence
        - runbookVersionId
        - runbookVersion
        - runbookKey
        - runbookName
        - parameters
        - riskLevel
        - expectedVerification
        - proposalHash
        - createdAt
      properties:
        id:
          type: string
          format: uuid
        diagnosisRunId:
          type: string
          format: uuid
        incidentVersion:
          type: integer
          format: int64
        summary:
          type: string
        hypotheses:
          type: array
          items:
            $ref: "#/components/schemas/Hypothesis"
        missingEvidence:
          type: array
          items:
            type: string
        runbookVersionId:
          type: [string, "null"]
          format: uuid
        runbookVersion:
          type: [integer, "null"]
        runbookKey:
          type: [string, "null"]
        runbookName:
          type: [string, "null"]
        parameters:
          type: object
          additionalProperties: true
        riskLevel:
          $ref: "#/components/schemas/RiskLevel"
        expectedVerification:
          oneOf:
            - $ref: "#/components/schemas/VerificationExpectation"
            - type: "null"
        proposalHash:
          type: string
        createdAt:
          type: string
          format: date-time
    EvidenceReference:
      type: object
      additionalProperties: false
      required: [id, sourceType, sourceRef, contentHash, capturedAt, truncated]
      properties:
        id:
          type: string
          format: uuid
        sourceType:
          type: string
        sourceRef:
          type: string
        contentHash:
          type: string
        capturedAt:
          type: string
          format: date-time
        truncated:
          type: boolean
    EvidenceSnapshot:
      type: object
      additionalProperties: false
      required:
        [id, sourceType, sourceRef, contentHash, capturedAt, truncated, redactedPayload]
      properties:
        id:
          type: string
          format: uuid
        sourceType:
          type: string
        sourceRef:
          type: string
        contentHash:
          type: string
        capturedAt:
          type: string
          format: date-time
        truncated:
          type: boolean
        redactedPayload:
          type: object
          additionalProperties: true
    CockpitApproval:
      type: object
      additionalProperties: false
      required:
        - id
        - proposalId
        - proposalHash
        - targetAlias
        - status
        - resourceVersion
        - requiredApprovals
        - approvals
        - rejections
        - requesterId
        - requesterSubject
        - requesterDisplayName
        - independentApproverRequired
        - createdAt
        - expiresAt
      properties:
        id:
          type: string
          format: uuid
        proposalId:
          type: string
          format: uuid
        proposalHash:
          type: string
        targetAlias:
          type: string
        status:
          $ref: "#/components/schemas/ApprovalStatus"
        resourceVersion:
          type: integer
          format: int64
        requiredApprovals:
          type: integer
        approvals:
          type: integer
        rejections:
          type: integer
        requesterId:
          type: string
          format: uuid
        requesterSubject:
          type: string
        requesterDisplayName:
          type: string
        independentApproverRequired:
          type: boolean
        createdAt:
          type: string
          format: date-time
        expiresAt:
          type: string
          format: date-time
        decidedAt:
          type: [string, "null"]
          format: date-time
    CockpitExecution:
      type: object
      additionalProperties: false
      required:
        - id
        - proposalId
        - approvalRequestId
        - status
        - targetAlias
        - fencingToken
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
        proposalId:
          type: string
          format: uuid
        approvalRequestId:
          type: string
          format: uuid
        status:
          $ref: "#/components/schemas/ExecutionStatus"
        targetAlias:
          type: string
        fencingToken:
          type: integer
          format: int64
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        startedAt:
          type: [string, "null"]
          format: date-time
        completedAt:
          type: [string, "null"]
          format: date-time
    IncidentCockpit:
      type: object
      additionalProperties: false
      required: [incident, diagnosis, evidence, activeApproval, latestExecution]
      properties:
        incident:
          $ref: "#/components/schemas/IncidentSummary"
        diagnosis:
          oneOf:
            - $ref: "#/components/schemas/CockpitDiagnosis"
            - type: "null"
        evidence:
          type: array
          maxItems: 100
          items:
            $ref: "#/components/schemas/EvidenceReference"
        activeApproval:
          oneOf:
            - $ref: "#/components/schemas/CockpitApproval"
            - type: "null"
        latestExecution:
          oneOf:
            - $ref: "#/components/schemas/CockpitExecution"
            - type: "null"
    IncidentTimelineItem:
      type: object
      additionalProperties: false
      required:
        [
          id,
          sequence,
          eventType,
          actorType,
          actorId,
          summary,
          traceId,
          evidenceIds,
          occurredAt,
        ]
      properties:
        id:
          type: string
          format: uuid
        sequence:
          type: integer
          format: int64
        eventType:
          type: string
        actorType:
          type: string
        actorId:
          type: string
        source:
          type: [string, "null"]
        sourceEventId:
          type: [string, "null"]
        summary:
          type: string
          description: Server-generated allowlisted summary; raw event payloads are never returned.
        traceId:
          type: [string, "null"]
          description: Validated distributed trace identifier when the producer supplied one.
        evidenceIds:
          type: array
          maxItems: 100
          items:
            type: string
            format: uuid
        occurredAt:
          type: string
          format: date-time
    IncidentTimelinePage:
      type: object
      additionalProperties: false
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/IncidentTimelineItem"
        nextCursor:
          type: [string, "null"]
...